refactor(rs): consolidate converters into navigation
